Updating protos of bitset/getfour.c

+++ ft_bitsetfour.c
@@ -6,13 +6,13 @@
 /*   By: kpiacent <kpiacent@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/03/05 18:28:04 by kpiacent          #+#    #+#             */
-/*   Updated: 2016/03/05 20:47:20 by kpiacent         ###   ########.fr       */
+/*   Updated: 2016/03/06 11:03:09 by kpiacent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bitsetfour(int *i, int val, size_t index)
+void	ft_bitsetfour(int *i, int val, unsigned int index)
 {
 	int		n;
